Move from u64 to u128 for sqrt_price and liquidity. This will provide more precision and it seems as if this is widely adopted in blockchain programs. Since rust does not have a built in u256 type, primitive types crate is used

+++ programs/pool/src/common/fixed_point_math.rs
@@ -11,13 +11,13 @@ const Q3232_FP_SHIFT: u8 = 32;
 /// fp<Q32.32>: floating point 32.32
 ///
 /// returns: u64
-pub fn mul_round_down_Q3232(ui: u64, fp: u64) -> Result<u64> {
+pub fn mul_round_down_Q3232(ui: u128, fp: u128) -> Result<u128> {
     if ui == 0 || fp == 0 {
         return Ok(0);
     }
     let product = ui
         .checked_mul(fp)
         .ok_or(SureError::MultiplictationQ3232Overflow)?;
-    let result = (product >> Q3232_FP_SHIFT) as u64;
+    let result = (product >> Q3232_FP_SHIFT);
     Ok(result)
 }
